+++ analysis/import/SCMLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/scmlogparser/input/metrics/MetricsFactory.kt
@@ -40,7 +40,9 @@ class MetricsFactory {
                 HighlyCoupledFiles(),
                 MedianCoupledFiles(),
                 AbsoluteCoupledChurn(),
-                AverageCodeChurnPerCommit()
+                AverageCodeChurnPerCommit(),
+                NumberOfRenames(),
+                AgeInWeeks()
         )
     }